requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
torch
